changed fileformat to be specific to vpc flow logs from hc
